backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djoser==2.2.3
Pillow==10.4.0
django-filter==24.3
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-dotenv==1.0.1
